python-backend-2/requirements.txt
llama-index==0.12.47
jedi==0.19.2
llama-index-embeddings-huggingface==0.5.5
python-dotenv==1.1.1
rapidfuzz==3.13.0
llama-index-llms-google-genai==0.2.3
sse-starlette==2.4.1
oracledb
gunicorn
uvicorn[standard]
fastapi
elasticsearch==8.11.0
google.generativeai
spacy
llama_index.llms.vertex
python-multipart